+++ CHANGELOG.md
@@ -6,6 +6,8 @@ What's changed since pre-release v0.1.0-B2012006:
 
 - New features:
   - Added `Get-AzDocTemplateFile` cmdlet to scan for template files within a path. [#16](https://github.com/Azure/PSDocs.Azure/issues/16)
+- General improvements:
+  - Add support for document localization and JSON snippet formatting. [#18](https://github.com/Azure/PSDocs.Azure/issues/18)
 - Engineering:
   - Bump PSDocs dependency to v0.7.0.